requirements.txt
pandas==1.5.3
python-json-logger==2.0.7
python-Levenshtein==0.21.1
rouge-score==0.1.2
scikit-learn==1.2.2
scipy==1.10.1
timm==0.4.9
torch==2.0.0
torch-cluster==1.6.1
torch-geometric==2.3.1
torch-scatter==2.1.1
torchaudio==2.0.0
torchelastic==0.2.2
torchtext==0.15.0
torchvision==0.15.0
transformers==4.30.1
